# doc/api.yaml
{
  "openapi": "3.0.0",
  "paths": {
...
    "/track": {
      "post": {
        "operationId": "TrackController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Data for creating a new track",
          "content": {
            "application/json": {
              "schema": {
                "example": {
                  "name": "Track name",
                  "artistId": "123e4567-e89b-12d3-a456-426614174000",
                  "albumId": "123e4567-e89b-12d3-a456-426614174000",
                  "duration": 240
                },
                "$ref": "#/components/schemas/CreateTrackDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "track successfully created",
            "schema": {
              "example": {
                "id": "123e4567-e89b-12d3-a456-426614174000",
                "name": "Track name",
                "artistId": "123e4567-e89b-12d3-a456-426614174000",
                "albumId": "123e4567-e89b-12d3-a456-426614174000",
                "duration": 240
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseTrackDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request. artistId is invalid (not uuid)"
          }
        },
        "summary": "Create track",
        "tags": [
          "track"
        ]
      },
      "get": {
        "operationId": "TrackController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "track found",
            "schema": {
              "example": {
                "name": "Track name",
                "artistId": "123e4567-e89b-12d3-a456-426614174000",
                "albumId": "123e4567-e89b-12d3-a456-426614174000",
                "duration": 240
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResponseTrackDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Get all track",
        "tags": [
          "track"
        ]
      }
    },
    "/track/{id}": {
      "get": {
        "operationId": "TrackController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID track id",
            "schema": {
              "format": "uuid",
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "track found",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "123e4567-e89b-12d3-a456-426614174000",
                    "name": "Track name",
                    "artistId": "123e4567-e89b-12d3-a456-426614174000",
                    "albumId": "123e4567-e89b-12d3-a456-426614174000",
                    "duration": 240
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request. artistId is invalid (not uuid)"
          },
          "404": {
            "description": "track not found"
          }
        },
        "summary": "Get track by id",
        "tags": [
          "track"
        ]
      },
      "put": {
        "operationId": "TrackController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID track id",
            "schema": {
              "format": "uuid",
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Data for updating an existing track",
          "content": {
            "application/json": {
              "schema": {
                "example": {
                  "name": "Track name",
                  "artistId": "123e4567-e89b-12d3-a456-426614174000",
                  "albumId": "123e4567-e89b-12d3-a456-426614174000",
                  "duration": 240
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "track successfully updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseTrackDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request. artistId is invalid (not uuid)"
          },
          "404": {
            "description": "track not found"
          }
        },
        "summary": "Update track",
        "tags": [
          "track"
        ]
      },
      "delete": {
        "operationId": "TrackController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID track id",
            "schema": {
              "format": "uuid",
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "track successfully deleted"
          },
          "400": {
            "description": "Bad request. artistId is invalid (not uuid)"
          },
          "404": {
            "description": "track not found"
          }
        },
        "summary": "Delete track by id",
        "tags": [
          "track"
        ]
      }
    },
...
    "/favs/track/{id}": {
      "post": {
        "operationId": "FavoritesController_addTracks",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID track id",
            "schema": {
              "format": "uuid",
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Add track to favorites",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseTrackDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request. id is invalid (not uuid)"
          },
          "422": {
            "description": "track not found"
          }
        },
        "summary": "Add track to favorites",
        "tags": [
          "favorites"
        ]
      },
      "delete": {
        "operationId": "FavoritesController_removeTracks",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID track id",
            "schema": {
              "format": "uuid",
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Remove track from favorites"
          },
          "400": {
            "description": "Bad request. id is invalid (not uuid)"
          },
          "422": {
            "description": "track not found"
          }
        },
        "summary": "Remove track from favorites",
        "tags": [
          "favorites"
        ]
      }
    }
  },
  "info": {
    "title": "Home Library Service",
    "description": "Home music library service",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateTrackDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "track name",
            "description": "track name"
          },
          "artistId": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "format": "uuid | null",
            "description": "Unique artist ID"
          },
          "albumId": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "format": "uuid | null",
            "description": "Unique album ID"
          },
          "duration": {
            "type": "number",
            "example": 240,
            "description": "track duration"
          }
        },
        "required": [
          "name",
          "artistId",
          "albumId",
          "duration"
        ]
      },
      "ResponseTrackDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique track ID",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "name": {
            "type": "string",
            "description": "Name of the track",
            "example": "Track name"
          },
          "artistId": {
            "type": "string",
            "description": "ID of the artist associated with the track",
            "example": "123e4567-e89b-12d3-a456-426614174000 | null"
          },
          "albumId": {
            "type": "string",
            "description": "ID of the album associated with the track",
            "example": "123e4567-e89b-12d3-a456-426614174000 | null"
          },
          "duration": {
            "type": "number",
            "description": "Duration of the track in seconds",
            "example": 240
          }
        },
        "required": [
          "id",
          "name",
          "artistId",
          "albumId",
          "duration"
        ]
      },
...
    }
  }
}